Kjør dev-profilen uten at man trenger maskinporten-token
